Packing modules using webpack.
