--- options/options.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-		<link rel="stylesheet" href="popup/popup.css">
-
-	</head>
-	<body>
-		<form id='optionsform'>
-			<label>Intheam API key</label>
-			<input type='text' id='intheamapikey' />
-			</br>
-			<label>Refresh Interval in Minutes</label>
-			<input type='number' id='refreshinterval' required min=1
-						 pattern='[0-9]+'/>
-			</br>
-			<input class='btn btn-primary' type='submit' value='Save' />
-		</form>
-		<button class="btn btn-secondary" id="syncnowbutton" >Sync Now</button>
-		</div>
-	</body>
-	<script src="/third_party/jquery.min.js"></script>
-	<script src="/third_party/bootstrap.min.js"></script>
-	<script src='/lib/sync.js'></script>
-	<script src='/options/options.js'></script>
-</html>
