--- src/content.ts
@@ -1,3 +1,0 @@
-export function test() {
-  chrome.runtime.sendMessage({}, response => {});
-}
